pneumonia_detection/requirements.txt
numpy
tensorflow
matplotlib
Pillow